feat: Ignore comments when comparing pages to see if page has changed
